Adding requested changes

+++ cogs/events.py
@@ -25,8 +25,9 @@ class Events(Cog):
             else:
                 embed = msg.embeds[0]
             if (
-                not embed.title == "Match Overview - SR Tournament Draft"
-                and not embed.description == "Game was found! Time to ready up!"
+                (not embed.title == "Match Overview - SR Tournament Draft")
+                and (not embed.description == "Game was found! Time to ready up!")
+                and (not embed.description == "Mentioned players have been removed from the queue for not being ready on time.")
             ):
                 try:
                     await msg.delete()
